project/requirements.txt
Django==3.0.7
django-crispy-forms==1.9.0
django-map-widgets==0.3.0
gunicorn==20.0.4
pre-commit==2.3.0
psycopg2-binary==2.8.5
python-dotenv==0.13.0
uvicorn==0.11.5
